requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.0